itester api changed

+++ src/main/java/com/bdcom/itester/api/ITesterAPI.java
@@ -22,7 +22,7 @@ public interface ITesterAPI {
 
     public int clearStatReliably(int socketId, int cardId, int portId);
 
-    public int setHeader(int socketId, int cardId, int portId, int validStreamCount, int length, byte[] strHead);
+    public int setHeader(int socketId, int cardId, int portId, int validStreamCount, int length, int[] strHead);
 
     public int setPayload(int socketId, int cardId, int portId, int length, byte[] data, int payloadType);
